api/django-docker/users/views.py
from django.contrib.auth import get_user_model
from rest_framework.mixins import CreateModelMixin, ListModelMixin
from rest_framework.viewsets import GenericViewSet

from .serializers import CreateUserSerializer, ListUserSerializer

User = get_user_model()


class UserViewSet(CreateModelMixin, ListModelMixin, GenericViewSet):
    queryset = User.objects.all()

    def get_serializer_class(self):
        if self.request.method == "POST":
            return CreateUserSerializer
        return ListUserSerializer
