api/pyproject.toml
[tool.poetry]
name = "django-docker"
version = "0.1.0"
description = ""
authors = ["Santi Garcia <santi.garcia@mahusai.global>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
gunicorn = "^20.1.0"
celery = "^5.3.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
redis = "^4.5.5"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
autopep8 = "^2.0.2"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
